+++ app/database/seeds/DatabaseSeeder.php
@@ -1,17 +1,18 @@
 <?php
 
 use Illuminate\Database\Seeder;
+use OpenDominion\Console\Commands\Game\DataSyncCommand;
 
 class DatabaseSeeder extends Seeder
 {
     /**
      * Run the database seeds.
-     *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
         $this->call(CoreDataSeeder::class);
+
+        Artisan::call(DataSyncCommand::class);
 
         if (app()->environment('local')) {
             $this->call(DevelopmentSeeder::class);