+++ src/Factories/DominionFactory.php
@@ -38,12 +38,12 @@ class DominionFactory
      * @param Round $round
      * @param Race $race
      * @param string $realmType Currently only 'random'. Future will support packs
-     * @param string $name
+     * @param string $rulerName
+     * @param string $dominionName
      *
-     * @throws RuntimeException
      * @return Dominion
      */
-    public function create(User $user, Round $round, Race $race, string $realmType, string $name): Dominion
+    public function create(User $user, Round $round, Race $race, string $realmType, string $rulerName, string $dominionName): Dominion
     {
         // todo: check if user already has a dominion in this round
         // todo: refactor $realmType into Realm $realm, generate new realm in RealmService from controller instead
@@ -72,7 +72,8 @@ class DominionFactory
             'realm_id' => $realm->id,
             'race_id' => $race->id,
 
-            'name' => $name,
+            'ruler_name' => $rulerName,
+            'name' => $dominionName,
             'prestige' => 250,
 
             'peasants' => 1300,