+++ tests/Feature/TickTest.php
@@ -3,7 +3,6 @@
 namespace OpenDominion\Tests\Feature;
 
 use Artisan;
-use CoreDataSeeder;
 use DB;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use OpenDominion\Calculators\Dominion\PopulationCalculator;
@@ -18,7 +17,7 @@ class TickTest extends AbstractBrowserKitTestCase
 
     public function testMoraleTick()
     {
-        $this->seed(CoreDataSeeder::class);
+        $this->seedDatabase();
         $user = $this->createUser();
         $round = $this->createRound();
         $dominion = $this->createDominion($user, $round);
@@ -37,7 +36,7 @@ class TickTest extends AbstractBrowserKitTestCase
 
     public function testQueuesTick()
     {
-        $this->seed(CoreDataSeeder::class);
+        $this->seedDatabase();
         $user = $this->createUser();
         $round = $this->createRound();
         $dominion = $this->createDominion($user, $round);
@@ -92,7 +91,7 @@ class TickTest extends AbstractBrowserKitTestCase
 
     public function testQueueShouldntTickLockedDominions()
     {
-        $this->seed(CoreDataSeeder::class);
+        $this->seedDatabase();
         $user = $this->createUser();
         $round = $this->createRound('-2 days', '-1 days');
         $dominion = $this->createDominion($user, $round);
@@ -140,7 +139,7 @@ class TickTest extends AbstractBrowserKitTestCase
 
     public function testResourcesGetGeneratedOnTheSameHourThatBuildingsComeIn()
     {
-        $this->seed(CoreDataSeeder::class);
+        $this->seedDatabase();
         $user = $this->createUser();
         $round = $this->createRound();
         $dominion = $this->createDominion($user, $round);
@@ -175,7 +174,7 @@ class TickTest extends AbstractBrowserKitTestCase
     // https://github.com/WaveHack/OpenDominion/issues/217
     public function testTheProperAmountOfPlatinumGetsAddedOnTick()
     {
-        $this->seed(CoreDataSeeder::class);
+        $this->seedDatabase();
         $user1 = $this->createUser();
         $user2 = $this->createUser();
         $round = $this->createRound();
@@ -244,7 +243,7 @@ class TickTest extends AbstractBrowserKitTestCase
     // https://github.com/WaveHack/OpenDominion/issues/227
     public function testTheProperAmountOfFoodGetsAddedOnTick()
     {
-        $this->seed(CoreDataSeeder::class);
+        $this->seedDatabase();
         $user = $this->createUser();
         $round = $this->createRound();
         $dominion = $this->createDominion($user, $round);